new alt table custom report, tidying up
