+++ frontend/src/pages/Policy/policyContent/CurrentEffortsContent.tsx
@@ -1,72 +1,123 @@
-export const cityEfforts = [
+export const economicResources = [
 	{
-		title: 'Atlanta Police Department (APD) Initiatives',
+		title: 'Atlanta Community Food Bank',
 		description: (
 			<>
-				The APD has typically been involved in strategies such as increased patrolling in high-crime areas, community policing efforts, and collaborations with federal law enforcement for gun crime investigations.
+				This organization helps address food insecurity, which is a critical
+				aspect of economic hardship. By providing access to basic needs, they
+				indirectly help in reducing stressors that can lead to violence.
 			</>
 		),
+		
 	},
 	{
-		title: 'Community-Based Programs',
+		title: 'WorkSource Atlanta',
 		description: (
 			<>
-				There have been initiatives that focus on community engagement and violence interruption, which often involve working with local organizations and community leaders to address the root causes of violence.
+				Offers job training and employment assistance, helping to bridge the gap
+				in economic opportunities and reduce unemployment, a key factor in
+				economic disparities.
 			</>
 		),
-	},
-	{
-		title: 'Legislation and Enforcement',
-		description: (
-			<>
-				There may be local and state laws that focus on gun control, such as background checks, restrictions on firearm types, and efforts to curb illegal gun trafficking.
-			</>
-		),
+		
 	},
 ]
 
-export const fultonCountyEfforts = [
+export const educationalResources = [
+  {
+    title: "Communities In Schools of Atlanta",
+    description: (
+      <>
+        This group works within local schools to provide resources and support,
+        ensuring that children stay in school and have access to quality
+        education and after-school programs.
+      </>
+    ),
+    link: "https://www.cisatlanta.org/",
+  },
+  {
+    title: "The Atlanta Educational Telecommunications Collaborative",
+    description: (
+      <>
+        Focuses on educational programming and initiatives, aiming to enrich
+        educational experiences in the community.
+      </>
+    ),
+    link: "https://www.wabe.org/",
+  },
+];
+
+export const justiceResources = [
 	{
-		title: 'Operation Shield',
+		title: 'The King Center',
 		description: (
 			<>
-				A joint effort between the APD and local businesses to increase surveillance and police presence.
+				Educates on the philosophy and methods of nonviolence and social change,
+				addressing racial injustice as a core element of reducing violence.
 			</>
 		),
+		link: 'https://thekingcenter.org/',
 	},
 	{
-		title: 'Cure Violence Program',
+		title: 'Southern Center for Human Rights',
 		description: (
 			<>
-				Focuses on treating violence as a public health issue, employing violence interrupters and outreach workers.
+				Works for equality, dignity, and justice for people impacted by the
+				criminal legal system in the South, including addressing issues that
+				lead to violence.
 			</>
 		),
-	},
-	{
-		title: 'Juvenile Diversion Programs',
-		description: (
-			<>
-				Aimed at reducing juvenile involvement in the criminal justice system, focusing on rehabilitation and community engagement.
-			</>
-		),
+		link: 'https://www.schr.org/',
 	},
 ]
 
-export const dekalbCountyEfforts = [
+export const mentalHealthResources = [
 	{
-		title: 'DeKalb County Gun Buyback Program',
+		title: 'NAMI Atlanta/Auburn',
 		description: (
 			<>
-				Encourages residents to turn in firearms in exchange for gift cards, aiming to reduce the number of guns on the street.
+				Offers resources, support, and education on mental health, helping to
+				destigmatize and provide critical mental health services in the
+				community.
 			</>
 		),
+		link: 'https://namiatlantaauburn.org/',
 	},
 	{
-		title: 'Community Policing Initiatives',
+		title: 'CHRIS 180',
 		description: (
 			<>
-				Focused on building trust between police and communities, including regular community meetings and police involvement in local events.
+				This organization focuses on healing and recovery from trauma, which is
+				particularly important in communities affected by gun violence.
 			</>
 		),
+		link: 'https://chris180.org/',
 	},
 ]
+
+export const communityResources = [
+	{
+		title: 'Cure Violence Atlanta',
+		description: (
+			<>
+				Works to stop the spread of violence in communities by using methods and
+				strategies associated with disease control – detecting and interrupting
+				conflicts, identifying and treating high-risk individuals, and changing
+				social norms.
+			</>
+		),
+		link: 'https://chris180.org/program/violence-prevention/',
+	},
+	{
+		title: 'Atlanta Police Foundation',
+		description: (
+			<>
+				While it's a law enforcement-related entity, they often engage in
+				community-based programs and partnerships to promote safety and prevent
+				violence.
+			</>
+		),
+		link: 'https://atlantapolicefoundation.org/',
+	},
+]
+
