+++ frontend/src/reports/CustomAltTable.tsx
@@ -1,0 +1,188 @@
+import React, { useRef } from 'react'
+import { Fips } from '../data/utils/Fips'
+import { Breakdowns } from '../data/query/Breakdowns'
+import { MetricQuery } from '../data/query/MetricQuery'
+import CardWrapper from '../cards/CardWrapper'
+import { DemographicGroup, ALL, TIME_PERIOD_LABEL, AGE } from '../data/utils/Constants'
+import type { DemographicType } from '../data/query/Breakdowns'
+import { DataTypeConfig, formatFieldValue, isPctType } from '../data/config/MetricConfig'
+import { splitIntoKnownsAndUnknowns } from '../data/utils/datasetutils'
+import { makeA11yTableData } from '../data/utils/DatasetTimeUtils'
+import { Row } from '../data/utils/DatasetTypes'
+import {
+  Table,
+  TableBody,
+  TableCell,
+  TableContainer,
+  TableHead,
+  TableRow,
+  Tooltip,
+} from '@mui/material'
+import WarningRoundedIcon from '@mui/icons-material/WarningRounded'
+import { DATA_CATALOG_PAGE_LINK } from '../utils/internalRoutes'
+
+interface CustomAltTableProps {
+  fips: Fips
+  dataTypeConfig: DataTypeConfig
+  demographicType: DemographicType
+  reportTitle: string
+  className?: string
+  selectedTableGroups?: DemographicGroup[]
+}
+
+export default function CustomAltTable(props: CustomAltTableProps) {
+  const metricConfigRates = props.dataTypeConfig.metrics?.per100k ?? props.dataTypeConfig.metrics?.pct_rate ?? props.dataTypeConfig.metrics?.index
+
+  const breakdowns = Breakdowns.forFips(props.fips).addBreakdown(
+    props.demographicType
+  )
+
+  const ratesQuery = new MetricQuery(
+    metricConfigRates.metricId,
+    breakdowns,
+    props.dataTypeConfig.dataTypeId,
+    'historical'
+  )
+
+  const queries = [ratesQuery]
+
+  return (
+    <CardWrapper
+      downloadTitle={props.reportTitle}
+      queries={queries}
+      minHeight={400}
+      reportTitle={props.reportTitle}
+      className={`rounded-sm relative m-2 p-3 shadow-raised bg-white ${props.className}`}
+    >
+      {([queryResponseRates]) => {
+        const ratesData = queryResponseRates.getValidRowsForField(
+          metricConfigRates.metricId
+        )
+
+        const [knownRatesData, unknownPctShareData] = splitIntoKnownsAndUnknowns(
+          ratesData,
+          props.demographicType
+        )
+
+        const accessibleData = makeA11yTableData(
+          knownRatesData as Row[],
+          unknownPctShareData as Row[],
+          props.demographicType,
+          metricConfigRates,
+          undefined,
+          props.selectedTableGroups ?? [ALL],
+          false
+        )
+
+        const latestTimePeriod: string = accessibleData[0][TIME_PERIOD_LABEL]
+        const earliestTimePeriod: string =
+          accessibleData[accessibleData.length - 1][TIME_PERIOD_LABEL]
+
+        const tableRef = useRef(null)
+        const linkRef = useRef(null)
+        const optionalAgesPrefix = props.demographicType === AGE ? 'Ages ' : ''
+
+        return (
+          <>
+            {/* Render the Table */}
+            <TableContainer className='flex max-h-sm caption-top self-center overflow-auto'>
+              <Table
+                tabIndex={0}
+                ref={tableRef}
+                className='m-3 w-98p whitespace-nowrap rounded-sm border border-altDark'
+                size='small'
+                stickyHeader
+              >
+
+                <TableHead>
+                  <TableRow>
+            
+                    {Object.keys(accessibleData[0]).map((key) => {
+                      const isTimeCol = key === TIME_PERIOD_LABEL
+                      const isUnknownPctCol = key.includes('with unknown ')
+
+                      const dataColumnLabel = metricConfigRates.shortLabel
+
+                      return (
+                        <TableCell
+                          key={key}
+                          style={{
+                            whiteSpace: 'normal',
+                            wordWrap: 'break-word',
+                          }}
+                          className='break-words border-0 border-b border-altDark bg-white leading-lhSomeSpace'
+                        >
+                          {!isTimeCol &&
+                            key !== ALL &&
+                            !isUnknownPctCol &&
+                            optionalAgesPrefix}
+                          {key.replaceAll('_', ' ')}
+                          {!isTimeCol &&
+                            !isUnknownPctCol &&
+                            ` ${dataColumnLabel}`}
+                          {isTimeCol &&
+                            ` (${earliestTimePeriod} - ${latestTimePeriod})`}
+                        </TableCell>
+                      )
+                    })}
+                  </TableRow>
+                </TableHead>
+
+                <TableBody>
+                  {accessibleData.map((row) => {
+                    const keys = Object.keys(row)
+                    const demographicGroup = row[props.demographicType] // Assumes demographicType is used as a key for the group
+                    return (
+                      <TableRow
+                        key={row[TIME_PERIOD_LABEL]}
+                        className='odd:bg-tableZebra even:bg-white'
+                      >
+          
+                        {keys.map((key) => {
+                          const isTimePeriod = key === TIME_PERIOD_LABEL
+
+                          const appendPct =
+                            key.includes('with unknown ') ||
+                            isPctType(metricConfigRates.type)
+                          return (
+                            <TableCell
+                              key={key}
+                              style={{
+                                whiteSpace: 'normal',
+                                wordWrap: 'break-word',
+                              }}
+                            >
+                              {row[key] == null ? (
+                                <>
+                                  <Tooltip title='Insufficient data'>
+                                    <WarningRoundedIcon />
+                                  </Tooltip>
+                                  <span className='sr-only'>
+                                    Insufficient data
+                                  </span>
+                                </>
+                              ) : (
+                                <>
+                                  {isTimePeriod ? row[key] : formatFieldValue(
+                                    metricConfigRates.type,
+                                    row[key],
+                                    !appendPct
+                                  )}
+                                </>
+                              )}
+                            </TableCell>
+                          )
+                        })}
+                      </TableRow>
+                    )
+                  })}
+                </TableBody>
+              </Table>
+            </TableContainer>
+           
+          </>
+        )
+      }}
+    </CardWrapper>
+  )
+}
