+++ frontend/src/utils/internalRoutes.ts
@@ -17,7 +17,7 @@ export const OLD_AGE_ADJUSTMENT_LINK = '/ageadjustment'
 export const GUN_VIOLENCE_POLICY = '/policy' + '/gun-violence'
 export const CRISIS_OVERVIEW_TAB = GUN_VIOLENCE_POLICY + '/crisis-overview'
 export const DATA_COLLECTION_TAB = GUN_VIOLENCE_POLICY + '/data-collection'
-export const ADDRESSING_INEQUITIES_TAB = GUN_VIOLENCE_POLICY + '/addressing-inequities'
+export const OUR_FINDINGS_TAB = GUN_VIOLENCE_POLICY + '/our-findings'
 export const CURRENT_EFFORTS_TAB = GUN_VIOLENCE_POLICY + '/current-efforts'
 export const REFORM_OPPORTUNITIES_TAB = GUN_VIOLENCE_POLICY + '/reform-opportunities'
 export const HOW_TO_USE_THE_DATA_TAB = GUN_VIOLENCE_POLICY + '/how-to-use-the-data'
