+++ frontend/tailwind.config.ts
@@ -18,7 +18,7 @@ export default {
       '3xl': '64px',
     },
     boxShadow: {
-      'none': 'none',
+      'none': `[none]`,
       'raised':
         'rgba(0, 0, 0, 0.2) 0px 5px 5px -3px, rgba(0, 0, 0, 0.14) 0px 8px 10px 1px, rgba(0, 0, 0, 0.12) 0px 3px 14px 2px',
       'raised-tighter':
